Initial abstraction of Trace Identifiers (api module only).
